Refactor approximate search evaluation pipeline and add unit tests

+++ src/llm_search_quality_evaluation/vector_search_doctor/approximate_search_evaluator/evaluation/results.py
@@ -11,7 +11,7 @@ RESULTS_FILENAME = "evaluation_results.json"
 
 def write_results(
     result: EvaluationResult,
-    run: dict[str, dict[str, float]],
+    run: dict[str, dict[str, int]],
     qrels: dict[str, dict[str, int]],
     meta: EvaluationMeta,
     output_destination: Path,
